chore(build): generate icon sprite in assets build
affects: @rei/cdr-assets, @rei/cdr-icon

remove brand ai tokens/icon scripts and move icon sprite logic to assets build step. Add sprite to
examples to make sure it renders ok
