+++ README.md
@@ -16,6 +16,8 @@ Clone the project.
 
 `npm install`
 
+`npm run start` (build the monorepo)
+
 ### Run
 
 `npm run dev`
@@ -126,88 +128,6 @@ __TODO:__ Notes about Vue, components
 - Write the component name prefixed/namespaced with `cdr-`.
 - Horizontal theming with css-modules, dynamically binding classes based on theme
 
-## Folder Structure
-
-%`<file or directory>` = generated by a process (so don't worry too much about them/what's in them)
-
-```
-backstop_data                   #Files related to visual regression testing. Mostly generated by backstop.
-|-- %bitmaps_reference          #Generated by backstop reference - contains reference backstop images
-|-- %bitmaps_test               #Generated by backstop test - contains images that are tested against reference images
-|-- casper_scripts              #Contains scripts that will be used by backstop for doing things like clicks, hovers, etc.
-|-- %html_report                #Generated report for backstop test results
-build                           #Contains scripts and configs for the webpack build system (most of this is from the vue webpack template)
-|-- build.js                    #Prod build script - the entry point for `npm run build`
-|-- check-versions.js           #Checks node/npm versions (this came with the vue webpack template so I’m not sure if it’s super useful)
-|-- dev-client.js               #For hot module reloading in dev
-|-- dev-server.js               #Config for express dev server
-|-- mainPost.conf.js            #Webpack loader config for processing .postcss files for release vs dev
-|-- utils.js                    #From vue webpack template to mostly automate which loaders need to be used since vue supports scss, stylus, postcss, etc.
-|-- vue-loader.conf.js          #Config object for vue webpack loader (vue-loader)
-|-- webpack.dev.conf.js         #Webpack config options specific to dev
-|-- webpack.prod.conf.js        #Webpack config options specific to build/prod
-|-- webpack.test.conf.js        #Webpack config options specific to unit/e2e test
-config                          #Configuration files used in the webpack build steps (most of this is from the vue webpack template)
-|-- dev.env.js                  #Setting NODE_ENV for dev
-|-- index.js                    #Used in build steps for configuration options specific to each build type
-|-- prod.env.js                 #Setting NODE_ENV for prod
-|-- release.env.js              #Setting NODE_ENV for release
-|-- test.env.js                 #Setting NODE_ENV for test
-%dist/                          #Results from `npm run build` end up here
-%dist-docs/                     #Results from `npm run build:docs` end up here
-docs/                           #Location for general documentation guidelines that will be used in the styleguide
-src/                            #Source files
-|-- assets/                     #For things that will be included like images, fonts, icons, etc. (check the vue webpack template docs)
-  |-- icons/                    #`npm run get-icons` stores results here then creates the sprite from them. Also available for use individually via CSS, etc.
-|-- components/                 #All things components
-  |-- <comp>
-    |-- __tests__/              #Unit tests for components
-      |-- <comp>.spec.js        #Unit test files -- karma looks for them with this pattern
-    |-- examples/               #Examples of the component used in the proving grounds for backstop testing
-    |-- <comp>.backstop.js      #Config files for backstop visual regression testing
-    |-- <comp>.md               #Component documentation file displayed in styleguide -- styleguide examples are added here
-    |-- <comp>.pcss             #Styles specific to the component which are imported into the `.vue` file
-    |-- <comp>.vue              #Single file component
-  |-- _index.js                 #All components are imported/exported here
-  |-- examples.js               #All component example are imported/exported here for use in App.vue and the proving grounds
-|-- css/                        #All things css
-  |-- directives/               #CSS files for directives
-  |-- generic/                  #Reset and normalize styles. Also general styles that are related to large layout.
-  |-- settings/                 #Values and variables that are consumed in other files.
-  |-- themes/                   #Might be used for vertical and/or horizontal themes in the future.
-  |-- main.postcss              #CSS files imported here for ordering
-|-- directives/                 #All things directives
-  |-- <direc>
-    |-- examples/               #Examples of the directive used in the proving grounds for backstop testing
-    |-- <direc>.backstop.js     #Config files for backstop visual regression testing
-    |-- <direc>.js              #Code for the directive
-|-- mixins/                     #Mixins that are shared in components
-|-- utils/                      #Helper js files for things like debounce
-|-- App.vue                     #The component/directive catalog for building and regression testing -- the proving grounds
-|-- dev.js                      #Webpack entry point for everything !release
-|-- main.js                     #Webpack entry point for release that builds delivery assets
-static/                         #For static assets. These get included in dist/ and release/ exactly as is -- check the vue webpack template docs
-styleguide/                     #Unique config options for the styleguide
-test/                           #All things testing for both unit and e2e
-|-- e2e/                        #Everything for nightwatch testing
-|-- unit/                       #Everything for karma testing
-utils/                          #Utility node scripts
-|-- brandai.js                  #Fetches values from brand.ai for use in css
-|-- icons.js                    #Fetches icons from brand.ai and outputs a sprite of all of them
-.babelrc                        #Babel es6 transpiling config
-.eslintignore                   #Self explanatory
-.eslintrc.js                    #Config for eslint
-.gitignore                      #Self explanatory
-.pullapprove.yml                #Config for github PR approval
-.stylelintignore                #Directories that won't be linted by stylelint
-.stylelintrc                    #Config for stylelint
-.travis.yml                     #Holding onto it from v1, will need to be updated
-backstop.js                     #Main config for backstop. Global settings like viewport sizes and paths are set here. 
-index.html                      #app.vue is bootstrapped here and is used for both dev and build -- all css and js files will be injected automatically
-package.json                    #Everything for npm
-postcss.config.js               #Config for postcss -- add new postcss plugins here
-styleguide.config.js            #Configuration for vue-styleguidist
-```
 ## Commits 
 This project is Commitizen friendly.
 when creating a pull request run `git cz` rather than `git commit` and follow the prompts.
