+++ package.json
@@ -8,8 +8,6 @@
   "private": true,
   "scripts": {
     "commitmsg": "commitlint -e $GIT_PARAMS",
-    "get-tokens": "node utils/brandai.js",
-    "get-icons": "node utils/icons.js",
     "dev": "node build/dev-server.js",
     "unit": "cross-env BABEL_ENV=test jest --no-cache",
     "e2e": "node test/e2e/runner.js",
@@ -55,7 +53,6 @@
     "css-loader": "^0.28.4",
     "cz-conventional-changelog": "^2.1.0",
     "cz-lerna-changelog": "^1.2.1",
-    "download": "^6.2.5",
     "eslint": "^4.18.1",
     "eslint-config-airbnb-base": "^12.0.0",
     "eslint-friendly-formatter": "^3.0.0",
@@ -135,7 +132,6 @@
     "stylelint-config-standard": "^18.1.0",
     "stylelint-processor-html": "^1.0.0",
     "stylelint-webpack-plugin": "^0.10.3",
-    "svg-sprite": "^1.3.7",
     "svgxuse": "^1.2.4",
     "tenon-node": "^0.4.1",
     "theo": "^7.0.1",
