+++ src/cdr-assets/build.js
@@ -5,6 +5,7 @@ const fs = require('fs-extra');
 const klaw = require('klaw');
 const through2 = require('through2');
 const path = require('path');
+const SVGSpriter = require('svg-sprite');
 const { exec } = require('child_process');
 
 // core-css variables
@@ -36,27 +37,56 @@ fs.copySync(staticDir, distDir, {
   },
 });
 
-// make icon index
+// make icon index and sprite
 const icons = {};
+
+const spriterConfig = {
+  dest: './static/',
+  mode: {
+    symbol: {
+      dest: '.',
+      sprite: './dist/cdr-icons.svg',
+    },
+  },
+};
+
+const spriter = new SVGSpriter(spriterConfig);
 
 const onlySVG = through2.obj(function filterSVG(item, enc, next) {
   if (!item.stats.isDirectory() && path.extname(item.path) === '.svg') this.push(item);
   next();
 });
 
+// walk all icons
 klaw(path.resolve(__dirname, '..', 'assets/icons/rei'))
+  // filter to .svg (as a precaution)
   .pipe(onlySVG)
   .on('data', (item) => {
     const name = path.basename(item.path).split('.')[0];
-    fs.readFile(item.path, 'utf8', (err, data) => {
+    const iconFilePath = item.path;
+    fs.readFile(iconFilePath, 'utf8', (err, data) => {
       if (err) throw err;
 
+      // add to icon object
       Object.assign(icons, { [`${name}`]: data });
+      // add to spriter
+      spriter.add(iconFilePath, null, data);
     });
   })
   .on('end', () => {
+    // output icon.json
     fs.writeFile('./dist/icons.json', JSON.stringify(icons, null, '\t'), 'utf8', (err) => {
       if (err) throw err;
       console.log(chalk.green('Compiled icons.json'));// eslint-disable-line no-console
+    });
+
+    // output icon sprite
+    spriter.compile((error, result) => {
+      Object.keys(result).forEach((mode) => {
+        Object.keys(result[mode]).forEach((resource) => {
+          fs.outputFileSync(result[mode][resource].path, result[mode][resource].contents);
+          console.log(chalk.green('Created icon sprite'));// eslint-disable-line no-console
+        });
+      });
     });
   });
