+++ src/cdr-assets/package.json
@@ -22,6 +22,7 @@
     "klaw": "^2.1.1",
     "pkg-ok": "^1.1.0",
     "postcss-cli": "^4.1.1",
+    "svg-sprite": "^1.3.7",
     "through2": "^2.0.3"
   }
 }
